Updated to use TCP protocol
