+++ RedPitayaSDK/inc/implementer.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 #include "control.h"
-#include "udp.h"
+#include "tcp.h"
 #include "processing.h"
 
 int stop;
